chore: update per review
