+++ src/writeData/subscriptions/components/StatusHeader.tsx
@@ -105,11 +105,11 @@ const StatusHeader: FC<Props> = ({currentSubscription, setStatus}) => {
           <Label
             id="tid"
             key="tkey"
-            name={`${bulletins.length} Message${
+            name={`${bulletins.length} Notification${
               bulletins.length === 1 ? '' : 's'
             }`}
             color={InfluxColors.Grey25}
-            description={`${bulletins.length} Message${
+            description={`${bulletins.length} Notification${
               bulletins.length === 1 ? '' : 's'
             }`}
             onClick={handleShowMessages}
